Added debug functions to draw tile and get map position with the mouse and function to move the player with mouse click

+++ src/Map.cpp
@@ -249,8 +249,8 @@ Vector2D Map::MapToWorld(int x, int y) const
 Vector2D Map::WorldToMap(int x, int y) {
 
     Vector2D ret(0, 0);
-    ret.setX(x / mapData.tileWidth);
-    ret.setY(y / mapData.tileHeight);
+    ret.setX((float)(x / mapData.tileWidth));
+    ret.setY((float)(y / mapData.tileHeight));
 
     return ret;
 }
@@ -283,7 +283,7 @@ Vector2D Map::GetMapSizeInPixels()
 
 Vector2D Map::GetMapSizeInTiles()
 {
-    return Vector2D(mapData.width, mapData.height);
+    return Vector2D((float)mapData.width, (float)mapData.height);
 }
 
 // Method to get the navigation layer from the map
